+++ backend/entityservice/tests/util.py
@@ -126,7 +126,7 @@ def get_project_description(requests, new_project_data):
 
 
 def create_project_no_data(requests,
-                           result_type='mapping', number_parties=None):
+                           result_type='groups', number_parties=None):
     number_parties_param = (
         {} if number_parties is None else {'number_parties': number_parties})
     new_project_response = requests.post(url + '/projects',
@@ -141,7 +141,7 @@ def create_project_no_data(requests,
 
 
 @contextmanager
-def temporary_blank_project(requests, result_type='mapping'):
+def temporary_blank_project(requests, result_type='groups'):
     # Code to acquire resource, e.g.:
     project = create_project_no_data(requests, result_type)
     yield project
@@ -152,7 +152,7 @@ def temporary_blank_project(requests, result_type='mapping'):
 def create_project_upload_fake_data(
         requests,
         sizes, overlap=0.75,
-        result_type='mapping', encoding_size=128):
+        result_type='groups', encoding_size=128):
     data = generate_overlapping_clk_data(
         sizes, overlap=overlap, encoding_size=encoding_size)
     new_project_data, json_responses = create_project_upload_data(
@@ -162,7 +162,7 @@ def create_project_upload_fake_data(
 
 
 def create_project_upload_data(
-        requests, data, result_type='mapping', binary=False, hash_size=None):
+        requests, data, result_type='groups', binary=False, hash_size=None):
     if binary and hash_size is None:
         raise ValueError('binary mode must specify a hash_size')
 
@@ -263,7 +263,6 @@ def wait_approx_run_time(size, assumed_rate=1_000_000):
     """
     number_comparisons = sum(
         size0 * size1 for size0, size1 in itertools.combinations(size, 2))
-    time.sleep(5)
     time.sleep(number_comparisons / assumed_rate)